refactor(Panel): ストーリーファイル内のBase参照をPanelに統一
- baseClassNameGenerator を panelClassNameGenerator に変更
- typeof Base を typeof Panel に変更
- Storybook タイトルを Components/Base から Components/Panel に変更
- インポートパスを Base から Panel に変更
